pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mame-curator"
version = "0.0.1"
description = "Curate the best playable arcade games from a MAME ROM set with a modern web UI."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Ants Schemel" }]
requires-python = ">=3.12"
keywords = ["mame", "retroarch", "rom", "arcade", "curation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment",
    "Typing :: Typed",
]

dependencies = [
    "lxml>=5.3",       # parser/ — DAT XML + listxml streaming
    "pydantic>=2.10",  # parser/ + filter/ + future api/ — validated models
    "pyyaml>=6.0",     # filter/ — overrides.yaml + sessions.yaml + config.yaml
    "rich>=13.9",      # cli/ — styled console output
]

# Per coding standards §8 "Phase-staged dependency declaration": runtime deps
# for not-yet-implemented phases live here until the importing code ships.
# Promote a dep into [project.dependencies] in the same commit as its first
# `import`. Until Phase 4 lands, end users running `mame-curator parse` should
# not pull the entire web stack just to count machines in a DAT.
[project.optional-dependencies]
api = [             # Phase 4 — FastAPI surface + SSE for live progress
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.28",
    "sse-starlette>=2.1",
]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25",
    "hypothesis>=6.122",
    "respx>=0.22",
    "ruff>=0.9",
    "mypy>=1.13",
    "bandit>=1.8",
    "pre-commit>=4.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
mame-curator = "mame_curator.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mame_curator"]

# ---------------------------------------------------------------------------
# Ruff: linter + formatter (replaces black/flake8/isort/pyupgrade)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
    "S",    # flake8-bandit (security)
    "D",    # pydocstyle
]
ignore = [
    "D100", # missing docstring in public module
    "D104", # missing docstring in public package
    "D203", # conflicts with D211
    "D213", # conflicts with D212
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "D", "RUF002", "RUF003"]
# FastAPI's canonical handler signature uses `param: T = Depends(...)` and
# `param: T = Body(...)` / `Query(...)` — calls in default args are the public
# idiom (see https://fastapi.tiangolo.com/tutorial/dependencies/), so B008 is
# silenced for the API routes only. Other modules still flag B008.
# D-class lints are also relaxed inside `api/`: the Pydantic-schema classes in
# `schemas.py` are pure type definitions whose names already convey intent,
# and route-handler functions are documented at the spec level
# (`docs/specs/P04.md` § Routes).
"src/mame_curator/api/routes/*.py" = ["B008", "D102", "D103"]
"src/mame_curator/api/schemas.py" = ["D101", "D106"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# mypy: strict type checking
# (Pinned to mypy initially; switch to Astral's Ty when 1.0 stable.)
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
no_implicit_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
files = ["src", "tests"]

# lxml ships partial stubs that conflict more often than they help; we treat its
# return types as Any at module boundaries and rely on test coverage instead.
[[tool.mypy.overrides]]
module = "lxml.*"
ignore_missing_imports = true

# hypothesis decorators (@given, @settings) read as untyped under strict mode;
# the decorators are type-erased on purpose. Tests that use them are still
# checked for body-level typing.
[[tool.mypy.overrides]]
module = [
    "tests.filter.test_property",
    "tests.copy.test_bios_properties",
    "tests.api.test_routes_config",
]
disallow_untyped_decorators = false

# FastAPI's @router.get / .post / .put / .patch / .delete decorators read as
# untyped under strict mode; they're the canonical idiom for declaring API
# routes. The decorators are type-erased on purpose (FastAPI introspects the
# handler signature itself); strict-mode `untyped-decorator` would force every
# handler to be silenced individually. Per-module override is the cleanest
# escape hatch.
[[tool.mypy.overrides]]
module = "mame_curator.api.routes.*"
disallow_untyped_decorators = false

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=mame_curator",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=85",  # backend overall — per-module floors enforced via per-phase acceptance (see standards §6)
]
asyncio_mode = "auto"

# ---------------------------------------------------------------------------
# coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["src/mame_curator"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
show_missing = true

# ---------------------------------------------------------------------------
# bandit: security linting
# ---------------------------------------------------------------------------
[tool.bandit]
exclude_dirs = ["tests", ".venv"]
